Refactor directories
The XFS Algorithms and Data Structures book says that XFS has five
directory types: "Short Form", "Block", "Leaf", "Node", and "BTree".
But its description isn't quite right.  There are really only three
types: "Short Form", "Block", and "Other", where "Other" directories can
have their data forks represented in two ways (extent lists or btrees),
and their Leaf entries stored in two ways also (single extents or btrees).

This refactor makes a new directory type: "Long Form" directories.  It
combines common code from Leaf, Node, and BTree dirs.  It also fixes the
previously unhandled case of a BTree directory whose Leaf was stored
in a single extent.

This also makes Dir3Block slightly more efficient by removing code from
Dir3::next's default implementation that Block dirs don't need.

Fixes #148
